CopyTool deletes with more care now

+++ varial/tools.py
@@ -107,6 +107,7 @@ class CopyTool(Tool):
             src = os.path.abspath(os.path.join(self.cwd, '..'))
         else:
             src = os.getcwd()
+        src_objs = glob.glob(src + '/*')
         dest = os.path.abspath(self.dest)
 
         # check for htaccess and copy it to src dirs
@@ -117,11 +118,13 @@ class CopyTool(Tool):
 
         # clean dest dir and copy
         if self.wipe_dest_dir:
+            src_basenames = list(os.path.basename(p) for p in src_objs)
             for f in glob.glob(dest + '/*'):
-                self.message('INFO Deleting: ' + f)
-                shutil.rmtree(f, True)
+                if os.path.isdir(f) and os.path.basename(f) in src_basenames:
+                    self.message('INFO Deleting: ' + f)
+                    shutil.rmtree(f, True)
         ign_pat = shutil.ignore_patterns(*self.ignore)
-        for f in glob.glob(src + '/*'):
+        for f in src_objs:
             if os.path.isdir(f):
                 f = os.path.basename(f)
                 shutil.copytree(
